requirements.txt
mss
pillow
opencv-python
pyautogui
pyobjc-core
pyobjc